requirements.txt
# --- Core API ---
fastapi==0.115.0
uvicorn==0.31.0
python-multipart==0.0.22
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2

# --- Security & Auth ---
python-jose[cryptography]==3.4.0
bcrypt==4.0.1
passlib==1.7.4

# --- Database & Cache ---
supabase==2.10.0

# --- Scrapers & Automation ---
playwright==1.49.0
playwright-stealth==1.0.6
2captcha-python==1.2.5

# --- Utils ---
pytz==2024.1
